Refactor LC3 code into project directory
