+++ src/projects/LC3/LC3.vue
