+++ src/projects/LC3/svgPlaceholder.html
